+++ dbdaora/hash/entity.py
@@ -4,4 +4,4 @@ from typing import Protocol
 class HashEntity(Protocol):
     @property
     def id(self) -> str:
-        ...
+        ...  # pragma: no cover