+++ dbdaora/sorted_set/repositories/datastore.py
@@ -7,6 +7,8 @@ from . import SortedSetRepository
 
 
 class DatastoreSortedSetRepository(SortedSetRepository[Key]):
+    __skip_cls_validation__ = ('DatastoreSortedSetRepository',)
+
     async def add_fallback(
         self, entity: Entity, *entities: Entity, **kwargs: Any
     ) -> None: